gungnir/requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
arrow==0.10.0
beautifulsoup4==4.6.0
billiard==3.3.0.23
celery==3.1.25
certifi==2017.4.17
chardet==3.0.3
cssselect==1.0.1
Django==1.11.1
django-celery==3.2.1
djangorestframework==3.6.3
feedfinder2==0.0.4
feedparser==5.2.1
idna==2.5
jieba3k==0.35.1
kombu==3.0.37
lxml==3.7.3
newspaper3k==0.2.2
nltk==3.2.4
numpy==1.12.1
olefile==0.44
packaging==16.8
pandas==0.20.1
Pillow==4.1.1
psycopg2==2.7.1
Pygments==2.2.0
pyparsing==2.2.0
python-dateutil==2.6.0
pytrends==4.1.1
pytz==2017.2
PyYAML==3.12
requests==2.17.3
requests-file==1.4.2
six==1.10.0
tldextract==2.1.0
urllib3==1.21.1
